requirements.txt
discord.py==1.7.3
python-dotenv==0.19.2
wit==6.0.0
youtube_dl==2021.12.17
